+++ app/Http/Controllers/cms/CountryController.php
@@ -150,7 +150,7 @@ class CountryController extends Controller
 
     public function countrytrashList()
     {
-        // $this->authorize('view', new Country());
+        $this->authorize('view', new Country());
         $data['trashedCountries'] = Country::onlyTrashed()->get();
  
         return  view('cms.trashed.countryTrashedList',$data);
@@ -159,14 +159,14 @@ class CountryController extends Controller
 
     public function restoreCountry($country_code)
     {
-        // $this->authorize('restore', new Country());
+        $this->authorize('restore', new Country());
         $country = Country::onlyTrashed()->where('country_code', $country_code)->restore();
         return back()->with('success','Successfully Restored');
 
     }
     public function forceDeleteCountry($country_code)
     {
-        // $this->authorize('forceDelete', new Country());
+        $this->authorize('forceDelete', new Country());
         $country = Country::onlyTrashed()->where('country_code', $country_code)->forceDelete();
  
         return back()->with('success','Permanently Deleted');
